Gracefully handle missing taskwarior
The taskwarior plugin creates an error message when task is not in the
$PATH variable. Users must manually clear the error before starting Vim.
This makes using the taskwarior plugin on platforms without taskwarior
cumbersome.

This change is a small plugin. The plugin sets the value of
g:loaded_taskwarrior to 1 when task is *not* present. This happens prior
to Vim loading bundle/vim-taskwarrior. As a result, the taskwarrior
plugin will never load. The advantage of this approach is that it does
not rely on changing the plugin.