Remove fzf.vim
This package is currently not configured correctly. Further, it relies
on the `fzf` program which is usually not in a default installation. So,
let's remove this package for now. We will reintroduce this package when
we are ready to test for `fzf`. We will also need to install both this
package and fzf-main.